Title: 컴퓨터와 함께하는 숫자 게임 2, Time: -1ms, Memory: -1MB, Status: Compile Error - CodeTree

+++ 240901/컴퓨터와 함께하는 숫자 게임 2/play-number-game-with-computer-2.java
@@ -16,7 +16,7 @@ public class Main {
         int minCnt = M + 1;
         int maxCnt = -1;
 
-        for(int x = a; x <=b; x++){
+        for(long x = a; x <=b; x++){
             int cnt = cntOfBinarySearch(x);
             minCnt = Math.min(minCnt, cnt);
             maxCnt = Math.max(maxCnt, cnt);
